Refactor: 폴더 정리 및 경로 수정
