+++ src/components/ChartItem/ChartItem.jsx
@@ -1,11 +1,11 @@
-import RoundImage from "components/Image/RoundImage/RoundImage";
+import Image from "components/Image";
 import styles from "components/ChartItem/ChartItem.module.scss";
 
 export default function ChartItem({ data }) {
   return (
     <div className={styles.chartItemContainer}>
       <div className={styles.content}>
-        <RoundImage size="sm"/>
+        <Image.Round size="sm"/>
         <span className={styles.ranking}>{data.rank}</span>
         <strong className={styles.name}>{data.name}</strong>
       </div>
