+++ src/components/ChartItem/ChartItem.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/mixin_var.scss";
+@import "styles/mixin_var.scss";
 
 .chartItemContainer {
   display: flex;
