--- src/components/Image/RoundImage/RoundImage.module.scss
@@ -1,26 +1,0 @@
-@import "../../../styles/mixin_var.scss";
-
-.roundImage {
-  border: 1px solid var(--brand-orange);
-  padding: 5px;
-  border-radius: 50%;
-  overflow: hidden;
-  &.sm {
-    width: 70px;
-    height: 70px;
-  }
-  &.md {
-    width: 100px;
-    height: 100px;
-  }
-  &.lg {
-    width: 128px;
-    height: 128px;
-  }
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border-radius: 50%;
-  }
-}
