+++ src/components/Image/RoundImage/index.js
@@ -1,0 +1,1 @@
+export { default } from './RoundImage';
