+++ src/components/Image/index.js
@@ -1,0 +1,9 @@
+import RectangleImage from "./RectangleImage";
+import RoundImage from "./RoundImage";
+
+const Image = {
+  Rectangle: RectangleImage,
+  Round: RoundImage,
+};
+
+export default Image;
